NTWrite CSort, NTWrite SerialBinning
NTWrite in CSort Improves Performance significantly. However it seems as
if we are doing the same work twice. will now try to bin directly in the
original function in a cached nontemporal manner.

Will again flip the queue structure so that we dont require to count the
number of elements while binning.

Added push_size as a generalization of pushing for memory.