STDP Experiment
In This I have added some code to change weights when processing the spikes that have arrived (anticausal rule). For this, the structure representing the synapse of the spike is a triplet of NStart(4-bit), NEnd(4-bit) and &Weight (8-bit). The nonsequential memory access trashes the timing. Uses same coding strategy as Cached Binning.

Earlier: 3.2Bil Spikes in 27s
Now: 3.2 Bil Spikes in 83s

If I remove the STDP, the timing halves to 39s (extra time bcuz of issues in converting synapse representation while storing spikes)

Need to do something about this but seems very hard. will need extremely efficient Radix sort and even then memory is not completely sequential.